+++ material/src/material.ts
@@ -1,0 +1,3 @@
+export {TitledDialogModule} from './modules/titledDialog.module';
+export {TitledDialogService} from './services/titledDialog/titledDialog.service';
+export {TitledDialogConfig, TITLED_DIALOG_DATA} from './misc/titledDialog.interface';