+++ structured-log/src/modules/consoleLog.module.ts
@@ -2,7 +2,7 @@ import {NgModule, ModuleWithProviders} from "@angular/core";
 import {CommonModule} from '@angular/common';
 
 import {ConsoleComponent} from '../components/console/console.component';
-import {CONSOLE_COMPONENT_SINK, STRUCTURED_LOG_LOGGER} from "../types/tokens";
+import {CONSOLE_COMPONENT_SINK, STRUCTURED_LOG_LOGGER, CONSOLE_COMPONENT_SINK_SERVICE_PROVIDER} from "../types/tokens";
 
 /**
  * Module containing component Console (log)
@@ -35,6 +35,7 @@ export class ConsoleLogModule
             ngModule: ConsoleLogModule,
             providers:
             [
+                CONSOLE_COMPONENT_SINK_SERVICE_PROVIDER,
                 CONSOLE_COMPONENT_SINK,
                 STRUCTURED_LOG_LOGGER
             ]